Perfected enslavement

+++ tests/moving/coding_wheels.h
@@ -1,8 +1,10 @@
 #ifndef CODING_WHEELS_H
 #define CODING_WHEELS_H
 
-extern volatile int tick_l;
-extern volatile int tick_r;
+#include <stdint.h>
+
+extern volatile uint16_t tick_l;
+extern volatile uint16_t tick_r;
 
 void coding_wheels_start(void);
